@@ -2,15 +2,15 @@ apply plugin: 'com.android.application'
 apply plugin: 'me.tatarka.retrolambda'
 
 android {
-  compileSdkVersion 25
+  compileSdkVersion 28
   buildToolsVersion "25.0.2"
   defaultConfig {
     applicationId "com.yuki312.androiddaggersample"
     minSdkVersion 19
-    targetSdkVersion 25
+    targetSdkVersion 28
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
   }
   buildTypes {
     release {
@@ -32,12 +32,12 @@ android {
 
 dependencies {
   compile fileTree(dir: 'libs', include: ['*.jar'])
-  androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+  androidTestCompile('androidx.test.espresso:espresso-core:3.1.0', {
     exclude group: 'com.android.support', module: 'support-annotations'
     exclude group: 'com.google.code.findbugs', module: 'jsr305'
   })
-  compile 'com.android.support:appcompat-v7:25.3.1'
-  compile 'com.android.support.constraint:constraint-layout:1.0.2'
+  compile 'androidx.appcompat:appcompat:1.0.0'
+  compile 'androidx.constraintlayout:constraintlayout:1.1.3'
   testCompile 'junit:junit:4.12'
 
   // Core dependencies
@@ -52,7 +52,7 @@ dependencies {
   compile 'com.google.dagger:dagger-android-support:2.11-rc1'
 
   // Architecture components
-  compile "android.arch.lifecycle:runtime:1.0.0-alpha1"
-  compile "android.arch.lifecycle:extensions:1.0.0-alpha1"
-  annotationProcessor "android.arch.lifecycle:compiler:1.0.0-alpha1"
+  compile 'androidx.lifecycle:lifecycle-runtime:2.0.0'
+  compile 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+  annotationProcessor 'androidx.lifecycle:lifecycle-compiler:2.0.0'
 }
